removed a clang tidy warning about implicit narrowing conversion
warning: narrowing conversion from 'size_t' (aka 'unsigned long long') to 'float' [bugprone-narrowing-conversions,cppcoreguidelines-narrowing-conversions]

+++ avogadro/core/molecule.cpp
@@ -554,7 +554,7 @@ Molecule::BondType Molecule::addBond(const AtomType& a, const AtomType& b,
 size_t calcNlogN(size_t n)
 {
   size_t aproxLog = 1;
-  float aux = n;
+  float aux = static_cast<float>(n);
   while (aux > 2.0f) {
     aux /= 2.0f;
     ++aproxLog;
